+++ src/Patches.cs
@@ -1,93 +1,130 @@
-﻿using GameNetcodeStuff;
-using HarmonyLib;
+﻿using System;
+using System.Linq;
 using UnityEngine;
+using MonoMod.Cil;
+using Mono.Cecil.Cil;
+using GameNetcodeStuff;
 
 namespace JetpackFallFix {
     class Patches {
-        static float timeSinceRoundStarted = 0;
-
-        [HarmonyPatch(typeof(StartOfRound), "Update")]
-        [HarmonyPostfix]
-        static void StartOfRound_Update_Postfix(ref StartOfRound __instance) {
-            // I don't know a better way to get timeSinceRoundStarted
-            timeSinceRoundStarted = __instance.timeSinceRoundStarted;
+        static void LogIfDebugBuild(string text){
+            #if DEBUG
+            Plugin.Logger.LogInfo(text);
+            #endif
         }
 
-        [HarmonyPatch(typeof(PlayerControllerB), "Awake")]
-        [HarmonyPostfix]
-        static void PlayerControllerB_Awake_Postfix(ref PlayerControllerB __instance) {
+        internal static void Init(){
+            On.GameNetcodeStuff.PlayerControllerB.Awake += PlayerControllerB_Awake;
+            IL.GameNetcodeStuff.PlayerControllerB.Update += PlayerControllerB_Update;
+            IL.GameNetcodeStuff.PlayerControllerB.PlayerHitGroundEffects += PlayerControllerB_PlayerHitGroundEffects;
+            On.GameNetcodeStuff.PlayerControllerB.PlayerHitGroundEffects += On_PlayerControllerB_PlayerHitGroundEffects;
+        }
+
+        private static void PlayerControllerB_Awake(On.GameNetcodeStuff.PlayerControllerB.orig_Awake orig, PlayerControllerB self)
+        {   
+            orig(self);
             // velocityAverageCount never gets reset, it only goes up when jetpack is used.
             // If this variable is greater than velocityMovingAverageLength, which is always 20,
             // averageVelocity will be higher when the jetpack is first used, which in turn means
             // you will take immediate fall damage when moving fast enough and activating the jetpack.
-            __instance.velocityAverageCount = 21;
+            self.velocityAverageCount = 21;
         }
 
-        [HarmonyPatch(typeof(PlayerControllerB), "DamagePlayer")]
-        [HarmonyPrefix]
-        static bool PlayerControllerB_DamagePlayer_Prefix(ref PlayerControllerB __instance, ref CauseOfDeath __3) {
-            //var myLogSource = BepInEx.Logging.Logger.CreateLogSource("JetpackFallFix");
-            if(__instance.jetpackControls && __3 == CauseOfDeath.Gravity) {
-                // Fix check for collision damage in air by adding originally missing argument QueryTriggerInteraction.Ignore
-                if(!Physics.CheckSphere(__instance.gameplayCamera.transform.position, 3f, StartOfRound.Instance.collidersAndRoomMaskAndDefault, QueryTriggerInteraction.Ignore)){
-                    //myLogSource.LogInfo($"Fix air damage");
+        private static void PlayerControllerB_Update(ILContext il)
+        {
+            /*
+            // Find:
+
+            if (Physics.CheckSphere(this.gameplayCamera.transform.position, 3f, StartOfRound.Instance.collidersAndRoomMaskAndDefault))
+            
+            // And add QueryTriggerInteraction.Ignore at the end of CheckSphere().
+            // This prevents us from colliding with trigger colliders when flying with the jetpack.
+            */
+
+            ILCursor c = new(il);
+            c.GotoNext(
+                x => x.MatchLdarg(0),
+                x => x.MatchLdfld<PlayerControllerB>("gameplayCamera"),
+                x => x.MatchCallvirt<Component>("get_transform"),
+                x => x.MatchCallvirt<Transform>("get_position"),
+                x => x.MatchLdcR4(3),
+                x => x.MatchCall<StartOfRound>("get_Instance"),
+                x => x.MatchLdfld<StartOfRound>("collidersAndRoomMaskAndDefault"),
+                x => x.MatchCall<Physics>("CheckSphere")
+            );
+            c.Index += 7;
+            c.Remove();
+            c.Emit(OpCodes.Ldc_I4_1);   // push QueryTriggerInteraction.Ignore to stack
+            c.Emit(                     // And call CheckSphere with QueryTriggerInteraction
+                OpCodes.Call, typeof(Physics)
+                .GetMethods()
+                .Where(x => x.Name == "CheckSphere")
+                .FirstOrDefault()
+            );
+            LogIfDebugBuild(il.ToString());
+        }
+
+        private static void On_PlayerControllerB_PlayerHitGroundEffects(On.GameNetcodeStuff.PlayerControllerB.orig_PlayerHitGroundEffects orig, PlayerControllerB self)
+        {
+            orig(self);
+            // We should reset averageVelocity when hitting ground.
+            self.averageVelocity = 0;
+        }
+
+        private static void PlayerControllerB_PlayerHitGroundEffects(ILContext il)
+        {
+            /*
+            // Find:
+
+            if (this.takingFallDamage
+                && !this.jetpackControls
+                && !this.disablingJetpackControls
+                && !this.isSpeedCheating
+            )
+
+            // And transform it to:
+
+            if (this.takingFallDamage
+                && [our Delegate method]
+                && !this.isSpeedCheating
+            )
+            */
+
+            ILCursor c = new(il);
+
+            c.GotoNext(
+                x => x.MatchLdarg(0),
+                x => x.MatchLdfld<PlayerControllerB>("takingFallDamage"),
+                x => x.MatchBrfalse(out _),
+                x => x.MatchLdarg(0),
+                x => x.MatchLdfld<PlayerControllerB>("jetpackControls"),
+                x => x.MatchBrtrue(out _),
+                x => x.MatchLdarg(0),
+                x => x.MatchLdfld<PlayerControllerB>("disablingJetpackControls"),
+                x => x.MatchBrtrue(out _),
+                x => x.MatchLdarg(0),
+                x => x.MatchLdfld<PlayerControllerB>("isSpeedCheating"),
+                x => x.MatchBrtrue(out _)
+            );
+            c.Index += 4;
+
+            // remove check for !jetpackControls && !disablingJetpackControls
+            c.RemoveRange(4); 
+            // above checks get replaced with this method
+            c.EmitDelegate<Func<PlayerControllerB, bool>>((self) =>
+            {
+                if(!(self.jetpackControls || self.disablingJetpackControls) || self.thisController.velocity.y < -15f){
+                    LogIfDebugBuild("Take fall damage!");
+                    // fallValueUncapped gets set to -8f when jetpack is used, so we do this.
+                    self.fallValueUncapped = self.thisController.velocity.y;
+                    // Also, fallValueUncapped gets reset back to -7 when we are on ground, in Update().
                     return false;
                 }
-                // Fix bug where you take damage when starting the jetpack in the ship while it is landing.
-                // Might already be fixed with setting velocityAverageCount to over 20, but  I'm keeping this just in case.
-                if(timeSinceRoundStarted < 10){
-                    Collider[] hitColliders = Physics.OverlapSphere(__instance.gameplayCamera.transform.position, 3f, StartOfRound.Instance.collidersAndRoomMaskAndDefault, QueryTriggerInteraction.Ignore);
-                    foreach (var hitCollider in hitColliders)
-                    {
-                        switch(hitCollider.name){
-                            case "ShipInside":
-                            case "ShipRails":
-                            case "WallInsulator":
-                            case "HangarDoorRight":
-                                //myLogSource.LogInfo($"Fix start damage");
-                                return false;
-                        }
-                    }
-                }
-            }
-            //myLogSource.LogInfo($"Took Damage, (vel): {__instance.thisController.velocity.magnitude}");
-            return true;
-        }
-
-        [HarmonyPatch(typeof(PlayerControllerB), "PlayerHitGroundEffects")]
-        [HarmonyPrefix]
-        static void PlayerControllerB_PlayerHitGroundEffects_Prefix(ref PlayerControllerB __instance) {
-            //var myLogSource = BepInEx.Logging.Logger.CreateLogSource("JetpackFallFix");
-            // New logic for jetpack falldamage that works more reliably. Without jetpack, this is essentially the same as original
-            // This is somewhat required to fix a bug where previous fall speed caused player to take damage when landing
-            if((__instance.jetpackControls || __instance.disablingJetpackControls) && __instance.fallValueUncapped >= -40){
-                if(__instance.thisController.velocity.y < -15){
-                    //myLogSource.LogInfo($"Jetpack fall damage, velocity.y: {__instance.thisController.velocity.y}, fallValueUncapped: {__instance.fallValueUncapped}");
-                    if (__instance.thisController.velocity.y < -45f) {
-                        __instance.DamagePlayer(__instance.thisController.velocity.y < -48.5f ? 100 : 80, hasDamageSFX: true, callRPC: true, CauseOfDeath.Gravity);
-                    }
-                    else {
-                        __instance.DamagePlayer(__instance.thisController.velocity.y < -40f ? 50 : 30, hasDamageSFX: true, callRPC: true, CauseOfDeath.Gravity);
-                    }
-                }
-            }
-            else if(__instance.takingFallDamage && !__instance.isSpeedCheating){
-                //myLogSource.LogInfo($"Basic fall damage, velocity.y: {__instance.thisController.velocity.y}, fallValueUncapped: {__instance.fallValueUncapped}");
-                if (__instance.fallValueUncapped < -45f) {
-                    __instance.DamagePlayer(__instance.fallValueUncapped < -48.5f ? 100 : 80, hasDamageSFX: true, callRPC: true, CauseOfDeath.Gravity);
-                }
-                else {
-                    __instance.DamagePlayer(__instance.fallValueUncapped < -40f ? 50 : 30, hasDamageSFX: true, callRPC: true, CauseOfDeath.Gravity);
-                }
-            }
-            else{
-                //myLogSource.LogInfo($"Touch ground, velocity.y: {__instance.thisController.velocity.y}, fallValueUncapped: {__instance.fallValueUncapped}");
-            }
-            // Reset averageVelocity to fix immediately taking damage when launching from ground
-            __instance.averageVelocity = 0f;
-            // Prevent orignal fall damage logic from running
-            __instance.takingFallDamage = false;
-            return;
+                LogIfDebugBuild("Prevented fall damage bug!");
+                // Not calling DamagePlayer() does not run this, so we do it here.
+                self.takingFallDamage = false;
+                return true;
+            });
         }
     }
 }