+++ src/Plugin.cs
@@ -1,16 +1,15 @@
-﻿// Nothing fun here, go to Patches.cs for some great content
-using BepInEx;
-using HarmonyLib;
+﻿using BepInEx;
+using BepInEx.Logging;
 
 namespace JetpackFallFix {
     [BepInPlugin(PluginInfo.PLUGIN_GUID, PluginInfo.PLUGIN_NAME, PluginInfo.PLUGIN_VERSION)]
-    public class JetpackFallFix : BaseUnityPlugin {
-        public static Harmony _harmony;
+    public class Plugin : BaseUnityPlugin {
+        internal static new ManualLogSource Logger;
         private void Awake() {
+            Logger = base.Logger;
             Logger.LogInfo($"Plugin {PluginInfo.PLUGIN_GUID} is loaded!");
 
-            _harmony = new Harmony(PluginInfo.PLUGIN_GUID);
-            _harmony.PatchAll(typeof(Patches));
+            Patches.Init();
         }
     }
 }